+++ src/RcsbFvStructure/RcsbFvStructure.tsx
@@ -1,14 +1,15 @@
 import * as React from "react";
 import {StructureViewerInterface} from "./StructureViewerInterface";
 import {RcsbFvDOMConstants} from "../RcsbFvConstants/RcsbFvConstants";
-import {RcsbFvSelectorManager} from "../RcsbFvSelection/RcsbFvSelectorManager";
+import {RcsbFvSelectorManager} from "../RcsbFvState/RcsbFvSelectorManager";
+import {RcsbFvStateManager} from "../RcsbFvState/RcsbFvStateManager";
 
 export interface RcsbFvStructureConfigInterface<R,S> {
     loadConfig: R | Array<R>;
     structureViewerConfig: S;
 }
 
-export class RcsbFvStructure<R,S> extends React.Component <RcsbFvStructureConfigInterface<R,S> & {structureViewer: StructureViewerInterface<R,S>, componentId: string, selectorManager: RcsbFvSelectorManager}, RcsbFvStructureConfigInterface<R,S> > {
+export class RcsbFvStructure<R,S> extends React.Component <RcsbFvStructureConfigInterface<R,S> & {structureViewer: StructureViewerInterface<R,S>, componentId: string,  stateManager: RcsbFvStateManager}, RcsbFvStructureConfigInterface<R,S> > {
 
     render():JSX.Element {
         return (
@@ -20,7 +21,7 @@ export class RcsbFvStructure<R,S> extends React.Component <RcsbFvStructureConfig
 
     async componentDidMount() {
         this.updateDimensions();
-        this.props.structureViewer.init(this.props.selectorManager, this.props.structureViewerConfig);
+        this.props.structureViewer.init(this.props.stateManager, this.props.structureViewerConfig);
         if(this.props.loadConfig)
             await this.props.structureViewer.load(this.props.loadConfig);
         window.addEventListener('resize', this.updateDimensions.bind(this));