+++ src/RcsbFvStructure/StructureViewers/MolstarViewer/MolstarManagerFactory.ts
@@ -4,11 +4,12 @@ import {Viewer, ViewerProps} from "@rcsb/rcsb-molstar/build/src/viewer";
 import {MolstarModelMapManager} from "./MolstarModelMapManager";
 import {MolstarCallbackManager} from "./MolstarCallbackManager";
 import {DataContainer} from "../../../Utils/DataContainer";
-import {RcsbFvSelectorManager} from "../../../RcsbFvSelection/RcsbFvSelectorManager";
+import {RcsbFvSelectorManager} from "../../../RcsbFvState/RcsbFvSelectorManager";
+import {RcsbFvStateManager} from "../../../RcsbFvState/RcsbFvStateManager";
 
 export class MolstarManagerFactory implements ViewerManagerFactoryInterface<LoadMolstarInterface,{viewerElement:string|HTMLElement,viewerProps:Partial<ViewerProps>}> {
 
-    public getViewerManagerFactory(selection: RcsbFvSelectorManager, viewerParams: {viewerElement: string | HTMLElement, viewerProps: Partial<ViewerProps>}) {
+    public getViewerManagerFactory(stateManager: RcsbFvStateManager, viewerParams: {viewerElement: string | HTMLElement, viewerProps: Partial<ViewerProps>}) {
         const loadingFlag: DataContainer<boolean> = new DataContainer(false);
         const innerSelectionFlag: DataContainer<boolean> = new DataContainer(false);
         const viewer = new Viewer(viewerParams.viewerElement, {
@@ -25,7 +26,7 @@ export class MolstarManagerFactory implements ViewerManagerFactoryInterface<Load
         const modelMapManager:MolstarModelMapManager = new MolstarModelMapManager(viewer);
         const callbackManager: MolstarCallbackManager =  new MolstarCallbackManager({
             viewer: viewer,
-            selection: selection,
+            stateManager: stateManager,
             loadingFlag: loadingFlag,
             modelMapManager: modelMapManager,
             innerSelectionFlag: innerSelectionFlag