+++ src/RcsbFvStructure/StructureViewers/MolstarViewer/StructureRepresentation.ts
@@ -8,8 +8,6 @@ import {StateObjectSelector} from "molstar/lib/mol-state";
 import {PluginStateObject} from "molstar/lib/mol-plugin-state/objects";
 import {StateObject} from "molstar/lib/mol-state/object";
 import {StateTransformer} from "molstar/lib/mol-state/transformer";
-import {Structure, StructureElement, StructureProperties as SP} from "molstar/lib/mol-model/structure";
-import {ChainInfo} from "../../StructureViewerInterface";
 
 type StructureObject = StateObjectSelector<PluginStateObject.Molecule.Structure, StateTransformer<StateObject<any, StateObject.Type<any>>, StateObject<any, StateObject.Type<any>>, any>>